# docs/openapi-generated.yaml
openapi: 3.1.0
info:
  title: 'Bar Assistant API'
...
  contact:
    name: 'Bar Assistant'
    url: 'https://barassistant.app'
    email: info@barassistant.app
  license:
    name: MIT
    url: 'https://github.com/karlomikus/bar-assistant/blob/master/LICENSE'
  version: '{{VERSION}}'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'Local docker development'
  -
    url: 'https://api.barassistant.app/api'
    description: Production
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: 'Authenticate user'
      description: 'Authenticate user and get auth token'
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Token' }
                type: object
        '400':
          description: 'Unable to authenticate'
      security: []
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register
      description: 'Register a new user'
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Profile' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
      security: []
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: 'Request password reset'
      description: 'Request a new password reset link'
      operationId: passwordForgot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  example: admin@example.com
              type: object
      responses:
        '204':
          description: 'Password reset link sent'
        '400':
          description: 'Unable to send password reset link'
      security: []
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: 'Reset password'
      description: 'Reset user password'
      operationId: passwordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                token:
                  type: string
                  example: token-from-email
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  minLength: 5
                  example: password
                password_confirmation:
                  type: string
                  minLength: 5
                  example: password
              type: object
      responses:
        '204':
          description: 'Password succssfully reset'
        '400':
          description: 'Unable to reset password'
      security: []
  '/auth/verify/{id}/{hash}':
    get:
      tags:
        - Authentication
      summary: 'Confirm account'
      description: 'Confirm user account, if applicable'
      operationId: confirmAccount
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a user'
          required: true
          schema:
            type: integer
        -
          name: hash
          in: path
          description: 'Hash string sent to user email'
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'Account confirmed'
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
      security: []
...
components:
  schemas:
...
    APIError:
      required:
        - type
        - message
      properties:
        type:
          type: string
          example: api_error
        message:
          type: string
          example: 'Resource record not found.'
      type: object
...
    BarMembership:
      required:
        - user_id
        - user_name
        - bar_id
        - is_shelf_public
        - use_parent_as_substitute
      properties:
        user_id:
          type: integer
          example: 1
        user_name:
          type: string
          example: Bartender
        bar_id:
          type: integer
          example: 1
        is_shelf_public:
          type: boolean
          example: true
        use_parent_as_substitute:
          type: boolean
          example: true
      type: object
...
    LoginRequest:
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: admin@example.com
        password:
          type: string
          format: password
          example: password
        token_name:
          type:
            - string
            - 'null'
          example: 'My device'
      type: object
...
    Profile:
      required:
        - id
        - name
        - cocktails_count
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Floral
        email:
          type: string
          example: example@example.com
        is_subscribed:
          type: boolean
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/BarMembership'
      type: object
...
    RegisterRequest:
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          example: admin@example.com
        name:
          type: string
          example: 'Bar Tender'
        password:
          type: string
          format: password
          minLength: 5
          example: password
      type: object
...
    Token:
      properties:
        token:
          type: string
          example: 1|dvWHLWuZbmWWFbjaUDla393Q9jK5Ou9ujWYPcvII
      type: object
...
  securitySchemes:
    user_token:
      type: http
      scheme: bearer
security:
  -
    user_token: []
...